pyproject.toml
[project]
name = "aslp"
version = "0.1.0"
description = "ASLP regression-resistant finetuning experiment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "torch==2.4.1",
  "transformers==4.46.2",
  "peft==0.13.2",
  "datasets==2.20.0",
  "accelerate>=1.1.0",
  "evaluate==0.4.3",
  "tqdm==4.66.5",
  "pyyaml==6.0.2",
  "numpy==1.26.4",
  "pytest==8.3.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
